requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asgiref=3.5.0=pypi_0
ca-certificates=2021.10.26=haa95532_4
certifi=2021.10.8=py37haa95532_2
dj-database-url=0.5.0=pypi_0
django=3.2.11=pypi_0
django-cors-headers=3.11.0=pypi_0
django-filter=1.1.0=pypi_0
django-heroku=0.3.1=pypi_0
djangorestframework=3.13.1=pypi_0
djangorestframework-filters=0.11.1=pypi_0
djangorestframework-jwt=1.11.0=pypi_0
djangorestframework-simplejwt=5.0.0=pypi_0
openssl=1.1.1m=h2bbff1b_0
pillow=9.0.0=pypi_0
pip=21.2.4=py37haa95532_0
psycopg2=2.9.3=pypi_0
pyjwt=2.3.0=pypi_0
python=3.7.11=h6244533_0
python-dotenv=0.19.2=pypi_0
pytz=2021.3=pypi_0
setuptools=58.0.4=py37haa95532_0
sqlite=3.37.0=h2bbff1b_0
sqlparse=0.4.2=pypi_0
typing-extensions=4.0.1=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.37.1=pyhd3eb1b0_0
whitenoise=5.3.0=pypi_0
wincertstore=0.2=py37haa95532_2
